vload/requirement.txt
azure-common==1.1.25
azure-core==1.4.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.3.0
